Examples/public-sdk-examples/03-ai-conversation/requirements.txt
bithuman>=1.6.0
livekit-agents[openai,silero]>=1.3
python-dotenv~=1.1
sounddevice>=0.4
